+++ vite.config.mjs
@@ -1,11 +1,20 @@
 import react from '@vitejs/plugin-react';
+import { codeInspectorPlugin } from 'code-inspector-plugin';
 import { defineConfig } from 'vite';
 import { createBlockletPlugin } from 'vite-plugin-blocklet';
 import svgr from 'vite-plugin-svgr';
 
-export default defineConfig(() => {
+export default defineConfig(({ mode }) => {
   return {
-    plugins: [react(), svgr(), createBlockletPlugin()],
+    plugins: [
+      react(),
+      svgr(),
+      createBlockletPlugin(),
+      mode === 'development' &&
+        codeInspectorPlugin({
+          bundler: 'vite',
+        }),
+    ],
     server: {
       fs: {
         strict: false,